pyproject.toml
[project]
name               = "spotify-actions"
version            = "0.1.0"
requires-python    = ">=3.13"
description        = "Schedule Actions on Spotify"
dependencies       = [
    "python-dotenv>=1.0.0",
    "spotipy>=2.0.0",
    "schedule>=1.0.0"
]

[build-system]
requires           = ["setuptools>=61.0","wheel"]
build-backend      = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "spotifyActionService/src"

[tool.setuptools.packages.find]
where              = ["spotifyActionService"]
include            = ["src*"]

[tool.ruff]
line-length        = 88
lint.extend-select = ["E", "F", "I", "B", "ANN", "UP", "TC"]
lint.ignore        = ["ANN101", "ANN102"]
fix                = true
fix-only           = false

[tool.pytest.ini_options]
minversion         = "6.0"
testpaths          = ["spotifyActionService/tst"]
python_files       = ["*Test.py"]

[tool.uv.sources]
spotify-actions = { workspace = true }

[dependency-groups]
dev = [
  "coverage>=7.8.0",
  "just>=0.8.162",
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1",
  "ruff>=0.11.8",
  "spotify-actions",
  "typing-extensions>=4.13.2",
]
